# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/admin/currencies": {
      "get": {
        "description": "Unlike the client-facing GET /currencies, this includes disabled currencies — managing them is the point of the screen.",
        "operationId": "AdminCurrenciesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CurrencyDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Every currency, including disabled ones, in operator order",
        "tags": [
          "admin"
        ]
      },
      "post": {
        "description": "The code is normalised to upper case and must be unique. Marking it default moves the flag off whatever holds it, atomically.",
        "operationId": "AdminCurrenciesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Add a currency",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/currencies/export": {
      "get": {
        "operationId": "AdminCurrenciesController_exportCurrencies",
        "parameters": [
          {
            "name": "format",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "csv"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A CSV file, named `currencies-<YYYY-MM-DD>.csv`.",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Export every currency as CSV, including disabled ones",
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/currencies/{code}": {
      "patch": {
        "description": "PATCH, not PUT: the code is the primary key and is referenced by wallets, transactions and transfers, so it is not editable — a rename would be a data migration across four money tables. Everything else is optional and only supplied fields change.",
        "operationId": "AdminCurrenciesController_update",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurrencyDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Update a currency",
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "description": "Refuses with 409 if any wallet exists in it — balances and append-only ledger history depend on the row. Disable it instead: that stops new wallets while keeping the existing ones readable and spendable.",
        "operationId": "AdminCurrenciesController_remove",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "summary": "Delete a currency that nobody holds",
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "OxShare CRM API",
    "description": "Forex/CFD Introducing-Broker CRM — Phase 1",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_portal_at",
        "description": "Portal session. Set by POST /v1/auth/login; httpOnly, so not settable here."
      },
      "admin": {
        "type": "apiKey",
        "in": "cookie",
        "name": "oxshare_crm_admin_at",
        "description": "Admin session. Set by POST /v1/admin/auth/login; httpOnly, so not settable here."
      }
    },
    "schemas": {
...
      "CurrencyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "USD"
          },
          "name": {
            "type": "string",
            "example": "US Dollar"
          },
          "symbol": {
            "type": "string",
            "example": "$"
          },
          "decimals": {
            "type": "number",
            "example": 2,
            "description": "Display precision only; storage is always 8dp."
          },
          "enabled": {
            "type": "boolean",
            "description": "Disabled currencies keep their wallets but accept no new ones."
          },
          "isDefault": {
            "type": "boolean",
            "description": "The currency a new client's first wallet opens in."
          },
          "sortOrder": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "code",
          "name",
          "symbol",
          "decimals",
          "enabled",
          "isDefault",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateCurrencyDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "EUR"
          },
          "name": {
            "type": "string",
            "example": "Euro"
          },
          "symbol": {
            "type": "string",
            "example": "€"
          },
          "decimals": {
            "type": "number",
            "default": 2,
            "minimum": 0,
            "maximum": 8
          },
          "enabled": {
            "type": "boolean",
            "default": true
          },
          "sortOrder": {
            "type": "number",
            "default": 0
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "code",
          "name",
          "symbol"
        ]
      },
      "UpdateCurrencyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Euro"
          },
          "symbol": {
            "type": "string",
            "example": "€"
          },
          "decimals": {
            "type": "number",
            "minimum": 0,
            "maximum": 8
          },
          "enabled": {
            "type": "boolean"
          },
          "sortOrder": {
            "type": "number"
          },
          "isDefault": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}